pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ru-fin-monitoring-helper"
version = "0.1.0"
description = "Helpers for RU financial monitoring / AML logging schemas and pipelines"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "run-as-daemon.ru"}]
dependencies = [
  "typer>=0.9.0",
  "PyYAML>=6.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "jsonschema>=4.21.0",
  "responses>=0.25.0",
  "ruff>=0.4.0",
  "mypy>=1.8.0",
  "yamllint>=1.35.0",
]

[project.scripts]
ru-fin-helper = "cli.main:app"

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
where = [
  ".",
  "ai/src",
  "sanctions/opensanctions_client/src",
  "sanctions/sanctions_network_client/src"
]
include = ["cli*", "ru_fin_helper*", "ru_fin_ai_helper*", "schemas*", "sanctions*", "ai*", "sanctions_network_client*"]
